Add files via upload
I just finished building a slider app using JavaScript, HTML, and CSS and I wanted to share it with the community. The app is fully responsive and works smoothly, I still have to make it work on all screen sizes. I will so the changes when I feel  more confortable understanding the screen sizes.